backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
motor==3.4.0 # Async MongoDB driver
simpy==4.0.2
pymongo==4.7.2 # Needed for BSON ObjectId handling if used directly sometimes, and potentially for export/import helper
python-dotenv==1.0.1
loguru==0.7.2 # Better logging